fix osquery restart process
cmd.Process.Kill() only kills the child process, but not those spawned by osqueryd.
When Restart() was called, osqueryd would die, but the extension process would remain, causing
a new osqueryd process to not start again. Killing the whole process group fixes the error.

Fixed the wait_healty test helper, which was shadowing the error and terminating early.
Added errgroup.Wait() in the restart method to wait for the old osquery extension to die.

Closes #221